# libs/GotIssues.Contracts/src/GotIssues.Contracts/Controllers/IssuesApi.cs
/*
 * Got Issues API
 *
 * A self-hosted, API-first issue and task tracker for internal engineering work.  This document is the product's user-facing documentation as well as its contract: every operation and schema is described for a reader who has never seen the implementation. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using GotIssues.Contracts.Attributes;
using GotIssues.Contracts.Models;

namespace GotIssues.Contracts.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class IssuesApiController : ControllerBase
    { 
        /// <summary>
        /// Create an issue in a project.
        /// </summary>
        /// <remarks>Creates an issue and allocates its number within the given project.  The number is allocated by the server and cannot be chosen.  **Requires a recognised role.** Any caller holding &#x60;admin&#x60; or &#x60;member&#x60; may create an issue; a token carrying neither receives 403. Unlike creating a project, this is not an administrative act.  A project that has exhausted its issue numbers — 999 999 999 of them, the most a key can express — is refused with 409. Issuing a key beyond that would produce an identifier this document&#39;s own pattern rejects. </remarks>
        /// <param name="projectKey">The key of the project the issue belongs to, for example &#x60;GOTI&#x60;.</param>
        /// <param name="createIssueRequest"></param>
        /// <response code="201">The issue was created.</response>
        /// <response code="400">The request was malformed or failed validation.</response>
        /// <response code="401">No credentials were supplied, or they were not valid.</response>
        /// <response code="403">The credentials were valid, but the caller&#39;s role does not permit this operation. Distinct from 401: the caller is known, and still refused. </response>
        /// <response code="404">The addressed resource does not exist — a project that was never created, or an issue key that corresponds to nothing. </response>
        /// <response code="409">The request conflicts with the state of the system — for a project, a key already in use; for an issue, a project that has exhausted its issue numbers. </response>
        /// <response code="500">The request could not be completed because of an unexpected failure.  Declared because the API can return it: an operation that reaches the database can fail in ways no validation anticipates, and a contract that lists only the outcomes it likes is as wrong as one that promises a body it does not send. The response is a problem document like every other failure — never an empty body, which is what a caller received before this was declared. </response>
        [HttpPost]
        [Route("/projects/{projectKey}/issues")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 201, type: typeof(Issue))]
        [ProducesResponseType(statusCode: 400, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 401, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 403, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 404, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 409, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 500, type: typeof(Problem))]
        public abstract Task<IActionResult> CreateIssue([FromRoute (Name = "projectKey")][Required][RegularExpression("^[A-Z][A-Z0-9]{1,9}$")][StringLength(10, MinimumLength=2)]string projectKey, [FromBody]CreateIssueRequest createIssueRequest);

        /// <summary>
        /// Read an issue by its key.
        /// </summary>
        /// <remarks>Returns the issue with the given key, for example &#x60;GOTI-1&#x60;.  **Requires a recognised role.** Any caller holding &#x60;admin&#x60; or &#x60;member&#x60; may read an issue; a token carrying neither receives 403. There is no per-project visibility — roles in this system are global.  Issues are addressed by the key people quote rather than by project and number separately, because that string is the thing written in commit messages and chat. </remarks>
        /// <param name="issueKey">The issue&#39;s key, for example &#x60;GOTI-1&#x60;.</param>
        /// <response code="200">The issue.</response>
        /// <response code="400">The request was malformed or failed validation.</response>
        /// <response code="401">No credentials were supplied, or they were not valid.</response>
        /// <response code="403">The credentials were valid, but the caller&#39;s role does not permit this operation. Distinct from 401: the caller is known, and still refused. </response>
        /// <response code="404">The addressed resource does not exist — a project that was never created, or an issue key that corresponds to nothing. </response>
        /// <response code="500">The request could not be completed because of an unexpected failure.  Declared because the API can return it: an operation that reaches the database can fail in ways no validation anticipates, and a contract that lists only the outcomes it likes is as wrong as one that promises a body it does not send. The response is a problem document like every other failure — never an empty body, which is what a caller received before this was declared. </response>
        [HttpGet]
        [Route("/issues/{issueKey}")]
        [Authorize]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(Issue))]
        [ProducesResponseType(statusCode: 400, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 401, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 403, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 404, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 500, type: typeof(Problem))]
        public abstract Task<IActionResult> GetIssue([FromRoute (Name = "issueKey")][Required][RegularExpression("^[A-Z][A-Z0-9]{1,9}-[1-9][0-9]{0,8}$")][StringLength(20, MinimumLength=4)]string issueKey);
    }
}