# libs/GotIssues.Contracts/src/GotIssues.Contracts/Controllers/ProjectsApi.cs
/*
 * Got Issues API
 *
 * A self-hosted, API-first issue and task tracker for internal engineering work.  This document is the product's user-facing documentation as well as its contract: every operation and schema is described for a reader who has never seen the implementation. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using GotIssues.Contracts.Attributes;
using GotIssues.Contracts.Models;

namespace GotIssues.Contracts.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class ProjectsApiController : ControllerBase
    { 
        /// <summary>
        /// Create a project.
        /// </summary>
        /// <remarks>Creates a project with a name and a key.  **Requires the &#x60;admin&#x60; role.** Creating a project is one of the three administrative acts in this system; a caller holding only &#x60;member&#x60; receives 403. The restriction is a property of the caller&#39;s role claim rather than of an OAuth scope, so it cannot be expressed in this document&#39;s security requirements — it is declared here, and in the 403 response, so that a client generating from this contract knows the endpoint can refuse an authenticated caller.  The key must be unique across the deployment. A key already in use is rejected with 409 rather than silently reused. </remarks>
        /// <param name="createProjectRequest"></param>
        /// <response code="201">The project was created.</response>
        /// <response code="400">The request was malformed or failed validation.</response>
        /// <response code="401">No credentials were supplied, or they were not valid.</response>
        /// <response code="403">The credentials were valid, but the caller&#39;s role does not permit this operation. Distinct from 401: the caller is known, and still refused. </response>
        /// <response code="409">The request conflicts with the state of the system — for a project, a key already in use; for an issue, a project that has exhausted its issue numbers. </response>
        /// <response code="500">The request could not be completed because of an unexpected failure.  Declared because the API can return it: an operation that reaches the database can fail in ways no validation anticipates, and a contract that lists only the outcomes it likes is as wrong as one that promises a body it does not send. The response is a problem document like every other failure — never an empty body, which is what a caller received before this was declared. </response>
        [HttpPost]
        [Route("/projects")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 201, type: typeof(Project))]
        [ProducesResponseType(statusCode: 400, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 401, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 403, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 409, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 500, type: typeof(Problem))]
        public abstract Task<IActionResult> CreateProject([FromBody]CreateProjectRequest createProjectRequest);

        /// <summary>
        /// List projects.
        /// </summary>
        /// <remarks>Returns a page of projects, newest first.  Any caller holding a recognised role may list projects; there is no per-project visibility, because roles in this system are global. </remarks>
        /// <param name="page">1-based page number. Out-of-range values are rejected with 400 rather than silently adjusted, matching pageSize.  Enforcement is the server&#39;s. Declaring a bound does not mean a generated client checks it before dispatching — the C# client generated from this document does not. </param>
        /// <param name="pageSize">Projects per page. The maximum is a declared constraint, so a larger value is rejected with 400 rather than silently reduced — a client asking for 10 000 and receiving 100 without being told is worse. </param>
        /// <response code="200">A page of projects.</response>
        /// <response code="400">The request was malformed or failed validation.</response>
        /// <response code="401">No credentials were supplied, or they were not valid.</response>
        /// <response code="403">The credentials were valid, but the caller&#39;s role does not permit this operation. Distinct from 401: the caller is known, and still refused. </response>
        /// <response code="500">The request could not be completed because of an unexpected failure.  Declared because the API can return it: an operation that reaches the database can fail in ways no validation anticipates, and a contract that lists only the outcomes it likes is as wrong as one that promises a body it does not send. The response is a problem document like every other failure — never an empty body, which is what a caller received before this was declared. </response>
        [HttpGet]
        [Route("/projects")]
        [Authorize]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(ProjectPage))]
        [ProducesResponseType(statusCode: 400, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 401, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 403, type: typeof(Problem))]
        [ProducesResponseType(statusCode: 500, type: typeof(Problem))]
        public abstract Task<IActionResult> ListProjects([FromQuery (Name = "page")][Range(1, 1000000)]int? page, [FromQuery (Name = "pageSize")][Range(1, 100)]int? pageSize);
    }
}