# spec/openapi.yaml
# The contract. Hand-authored, first — nothing here is generated FROM code.
#
# Server contracts and clients are generated from this document into libs/ by
# tools/generate.sh and are never hand-edited (ADR-0004). Operational endpoints
# (/health and friends) are deliberately absent: they serve operators, not clients
# generating against this document (ADR-0005).
openapi: 3.1.0

info:
  title: Got Issues API
  version: 0.1.0
  description: |
    A self-hosted, API-first issue and task tracker for internal engineering work.

    This document is the product's user-facing documentation as well as its
    contract: every operation and schema is described for a reader who has never
    seen the implementation.

servers:
  - url: http://localhost:8080
    description: Local development stack (docker compose)

security:
  - bearerAuth: []

tags:
  - name: Issues
    description: |
      A unit of work inside a project.

      Every issue carries a **key** of the form `GOTI-1`: the project's key, a
      hyphen, and a number allocated within that project. Numbering starts at 1 in
      each project and is independent between them, so `GOTI-1` and `PROJ-1` both
      exist and refer to different work.

      A number is never reused. If an issue is ever removed, its number stays
      retired rather than being handed to something else — `GOTI-7` pointing at
      different work than it did last week would be worse than a gap.
  - name: Projects
    description: |
      A project groups related work and owns a **key** — a short uppercase
      identifier such as `GOTI` that becomes the human-quotable half of every
      issue reference (`GOTI-123`).

      The key is chosen at creation and cannot be changed afterwards. Every issue
      reference derives from it, including ones written down outside this system,
      so renaming a key would orphan them all.

paths:
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List projects.
      description: |
        Returns a page of projects, newest first.

        Any caller holding a recognised role may list projects; there is no
        per-project visibility, because roles in this system are global.
      parameters:
        - name: page
          in: query
          required: false
          description: |
            1-based page number. Out-of-range values are rejected with 400 rather
            than silently adjusted, matching pageSize.

            Enforcement is the server's. Declaring a bound does not mean a generated
            client checks it before dispatching — the C# client generated from this
            document does not.
          schema:
            type: integer
            minimum: 1
            maximum: 1000000
            default: 1
        - name: pageSize
          in: query
          required: false
          description: |
            Projects per page. The maximum is a declared constraint, so a larger
            value is rejected with 400 rather than silently reduced — a client
            asking for 10 000 and receiving 100 without being told is worse.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of projects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a project.
      description: |
        Creates a project with a name and a key.

        **Requires the `admin` role.** Creating a project is one of the three
        administrative acts in this system; a caller holding only `member` receives
        403. The restriction is a property of the caller's role claim rather than of
        an OAuth scope, so it cannot be expressed in this document's security
        requirements — it is declared here, and in the 403 response, so that a client
        generating from this contract knows the endpoint can refuse an authenticated
        caller.

        The key must be unique across the deployment. A key already in use is
        rejected with 409 rather than silently reused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: The project was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /projects/{projectKey}/issues:
    post:
      operationId: createIssue
      tags: [Issues]
      summary: Create an issue in a project.
      description: |
        Creates an issue and allocates its number within the given project.

        The number is allocated by the server and cannot be chosen.

        **Requires a recognised role.** Any caller holding `admin` or `member` may
        create an issue; a token carrying neither receives 403. Unlike creating a
        project, this is not an administrative act.

        A project that has exhausted its issue numbers — 999 999 999 of them, the
        most a key can express — is refused with 409. Issuing a key beyond that would
        produce an identifier this document's own pattern rejects.
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project the issue belongs to, for example `GOTI`.
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9]{1,9}$'
            minLength: 2
            maxLength: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: The issue was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /issues/{issueKey}:
    get:
      operationId: getIssue
      tags: [Issues]
      summary: Read an issue by its key.
      description: |
        Returns the issue with the given key, for example `GOTI-1`.

        **Requires a recognised role.** Any caller holding `admin` or `member` may
        read an issue; a token carrying neither receives 403. There is no per-project
        visibility — roles in this system are global.

        Issues are addressed by the key people quote rather than by project and
        number separately, because that string is the thing written in commit
        messages and chat.
      parameters:
        - name: issueKey
          in: path
          required: true
          description: The issue's key, for example `GOTI-1`.
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9]{1,9}-[1-9][0-9]{0,8}$'
            minLength: 4
            maxLength: 20
      responses:
        '200':
          description: The issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT issued by the identity host. Machine clients obtain one through the
        client-credentials flow; the token carries a `role` claim the API reads
        per request and never stores.

  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: No credentials were supplied, or they were not valid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        The credentials were valid, but the caller's role does not permit this
        operation. Distinct from 401: the caller is known, and still refused.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: |
        The addressed resource does not exist — a project that was never created,
        or an issue key that corresponds to nothing.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: |
        The request could not be completed because of an unexpected failure.

        Declared because the API can return it: an operation that reaches the
        database can fail in ways no validation anticipates, and a contract that
        lists only the outcomes it likes is as wrong as one that promises a body it
        does not send. The response is a problem document like every other failure —
        never an empty body, which is what a caller received before this was
        declared.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: |
        The request conflicts with the state of the system — for a project, a key
        already in use; for an issue, a project that has exhausted its issue
        numbers.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Problem:
      type: object
      description: |
        An RFC 9457 problem document. Every failure in this API uses this shape,
        so clients get one error type generated rather than guessing per endpoint.
      properties:
        type:
          type: string
          description: A URI identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: A URI identifying this specific occurrence.

    Issue:
      type: object
      description: A unit of work inside a project.
      required: [id, key, projectKey, number, title, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: The issue's identifier.
        key:
          type: string
          pattern: '^[A-Z][A-Z0-9]{1,9}-[1-9][0-9]{0,8}$'
          description: |
            The issue's key: the project's key, a hyphen, and the number allocated
            within that project. This is the identifier people quote.
        projectKey:
          type: string
          pattern: '^[A-Z][A-Z0-9]{1,9}$'
          description: The key of the project this issue belongs to.
        number:
          type: integer
          minimum: 1
          maximum: 999999999
          description: |
            The issue's number within its project. Allocated by the server, starting
            at 1 in each project, and never reused.

            The maximum is not arbitrary: it is the largest number `key` can express,
            since that pattern allows nine digits. Without it the two fields could
            disagree — a number of ten digits would produce a key violating the very
            pattern this document declares, and the issue would be unreadable through
            the only operation that fetches one. A project reaching this limit is
            refused with 409 rather than issued a key it cannot use.
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^\u0000-\u001F\u007F]+$'
          description: |
            A one-line summary. C0 control characters and DEL are excluded, for the
            same two reasons a project name excludes them: PostgreSQL cannot store
            `U+0000` at all, and the ASCII line breaks and tabs in that range have no
            place in a title.

            Deliberately **not** full Unicode line-break handling: `U+0085` and
            `U+2028` are accepted. Refinement asked this ticket to decide rather than
            inherit the limit, and the decision is to keep the constraint the narrow,
            checkable one — a title carrying an exotic separator is a cosmetic
            problem, while a title carrying `U+0000` cannot be stored at all.
        description:
          type: [string, 'null']
          maxLength: 10000
          pattern: '^[^\u0000]*$'
          description: |
            Optional free text. Multi-line **by design** — unlike a title, this is
            where the reasoning goes, so line breaks are permitted and only `U+0000`
            is excluded, because PostgreSQL cannot store it in any text column.

            That difference is deliberate: the constraint on a title is about being
            one line, the constraint here is about being storable, and they are not
            the same rule (recorded on T-0005 from T-0004's review).
        createdAt:
          type: string
          format: date-time
          description: When the issue was created, in UTC.

    CreateIssueRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^\u0000-\u001F\u007F]+$'
          description: |
            A one-line summary. C0 control characters and DEL are excluded — see
            `Issue.title` for both reasons and for the limit this constraint
            deliberately stops at.

            `U+0085` and `U+2028` are accepted: the constraint is the narrow,
            checkable one.
        description:
          type: [string, 'null']
          maxLength: 10000
          pattern: '^[^\u0000]*$'
          description: |
            Optional free text; multi-line is expected. Explicitly null means the
            same as omitting it.

    Project:
      type: object
      description: A project, which groups related work and owns the key its issues are numbered under.
      required: [id, key, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: The project's identifier.
        key:
          type: string
          pattern: '^[A-Z][A-Z0-9]{1,9}$'
          minLength: 2
          maxLength: 10
          description: |
            The project's key: 2-10 characters, uppercase letters and digits,
            starting with a letter. Unique across the deployment and immutable
            once set.
        name:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^\u0000-\u001F\u007F]+$'
          description: |
            The project's display name. Names need not be unique — the key is the
            identifier, and requiring unique names would be a constraint on people
            rather than on data.

            C0 control characters and DEL are excluded. Two reasons, and they are
            not the same reason: PostgreSQL cannot store `U+0000` in text at all, so
            a name carrying one fails as an unhandled error rather than as
            validation; and the rest of the range carries tabs and line breaks, which
            a display name has no use for.

            This is not full Unicode line-break normalisation — `U+0085` and `U+2028`
            are accepted. The constraint is deliberately the narrow, checkable one.
          examples: [Got Issues]
        createdAt:
          type: string
          format: date-time
          description: When the project was created, in UTC.

    CreateProjectRequest:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          pattern: '^[A-Z][A-Z0-9]{1,9}$'
          minLength: 2
          maxLength: 10
          description: |
            The key this project's issues are numbered under, for example `GOTI`.
            Cannot be changed afterwards, so it is worth choosing deliberately.
        name:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^\u0000-\u001F\u007F]+$'
          description: |
            The project's display name. C0 control characters and DEL are excluded:
            `U+0000` cannot be stored by PostgreSQL at all, and the rest of the range
            carries tabs and line breaks a display name has no use for.

    ProjectPage:
      type: object
      description: |
        One page of results. Collection endpoints are always paginated — an
        unbounded result set is forbidden by the project's engineering standards.
      required: [items, page, pageSize]
      properties:
        items:
          type: array
          description: The projects on this page.
          items:
            $ref: '#/components/schemas/Project'
        page:
          type: integer
          description: The 1-based page number this response represents.
        pageSize:
          type: integer
          description: How many projects this page can hold.
        totalCount:
          type: integer
          description: Total projects matching the query, across all pages.